// src/calculation/calculation-methods.enum.ts
export enum CalculationMethodsEnum {
  // Uso de Recursos
  USAGE_EFFICIENCY = 'usage_efficiency', // Aumentar la eficiencia en el uso de suministros
  WASTE_REDUCTION = 'waste_reduction', // Reducir el desperdicio de suministros
  WASTE_MAINTENANCE = 'waste_maintenance', // Mantener el uso eficiente de los suministros

  // Eficiencia en el Cumplimiento de Tareas
  ON_TIME_COMPLETION = 'on_time_completion', // Aumentar el porcentaje de tareas completadas a tiempo
  ON_TIME_MAINTENANCE = 'on_time_maintenance', // Mantener el porcentaje de tareas completadas a tiempo

  // Gestión del Tiempo
  TIME_PRODUCTIVITY = 'time_productivity', // Aumentar el tiempo productivo
  IDLE_TIME_REDUCTION = 'idle_time_reduction', // Reducir el tiempo de inactividad
  TIME_MAINTENANCE = 'time_maintenance', // Mantener el tiempo utilizado en tareas

  // Calidad de Ejecución
  QUALITY_IMPROVEMENT = 'quality_improvement', // Aumentar la calidad en la ejecución de tareas
  QUALITY_MAINTENANCE = 'quality_maintenance', // Mantener la calidad de las tareas realizadas

  // Productividad por Trabajador
  PRODUCTIVITY_INCREASE = 'productivity_increase', // Aumentar la productividad por trabajador
  PRODUCTIVITY_MAINTENANCE = 'productivity_maintenance', // Mantener la productividad de los trabajadores

  // Cumplimiento de Plazos
  DEADLINE_COMPLIANCE = 'deadline_compliance', // Aumentar el cumplimiento de plazos
  DEADLINE_MAINTENANCE = 'deadline_maintenance', // Mantener el cumplimiento de plazos

  // Desempeño General por Tipo de Actividad
  ACTIVITY_PERFORMANCE_INCREASE = 'activity_performance_increase', // Aumentar el desempeño por tipo de actividad
  ACTIVITY_PERFORMANCE_MAINTENANCE = 'activity_performance_maintenance', // Mantener el desempeño por tipo de actividad
}